working unit test for moving to existing media group #26

+++ src/Tests/Commands/MoveFilesToMediaLibraryCommandTest.cs
@@ -19,7 +19,11 @@ public class MoveFilesToMediaLibraryCommandTest
         FileMovedToMediaLibraryEventArgs? commandEventArgs = null; // Hilfsvariable für das Erfassen der Event-Informationen
         command.FileMovedToMediaLibrary += (sender, e) => commandEventArgs = e;
         var expectedNewMediaGroupPath = new FileInfo(Path.Combine(command.MediaLibraryPath, "2023-03-14 Primeli Kurzaufnahme/2023-03-14 Primeli Kurzaufnahme.m4v"));
-        File.Delete(expectedNewMediaGroupPath.FullName); // Sicherstellen, dass die Datei gelöscht ist
+
+        if (File.Exists(expectedNewMediaGroupPath.FullName))
+        {
+            File.Delete(expectedNewMediaGroupPath.FullName); // Sicherstellen, dass die Datei gelöscht ist
+        }
 
         // Act
         var result = command.Execute();
@@ -31,9 +35,45 @@ public class MoveFilesToMediaLibraryCommandTest
         Assert.IsTrue(commandEventArgs.FileInfo.Exists);
         Assert.IsFalse(commandEventArgs.HasMovedToExistingMediaGroup); // Die Datei wurde nicht in eine bestehende Mediengruppe verschoben
         Assert.IsFalse(commandEventArgs.HasTargetFileBeenOverwritten); // Die Datei wurde nicht überschrieben
-        Assert.IsTrue(File.Exists(expectedNewMediaGroupPath.FullName));
-        
-        // Clean up
-        File.Delete(expectedNewMediaGroupPath.FullName);
+    }
+
+    [TestMethod] // Wird korrekt in eine bestehende Mediengruppe verschoben 
+    public void Execute_ShouldMoveFileToExistingMediaGroup_WhenFileIsAssignedToExistingMediaGroup()
+    {
+        // Arrange
+        var sourceFile = new FileInfo("Data/Input/Testcase 5/2023-06-15 Balkenmäher.m4v");
+        var mediaLibraryDirectory = new DirectoryInfo("Data/Output/Mediathek");
+        var existingMediaGroupDirectory = new DirectoryInfo(Path.Combine(mediaLibraryDirectory.FullName, "Familie/2023/"));
+        var fileInExistingMediaGroupDirectory = new FileInfo(Path.Combine(existingMediaGroupDirectory.FullName, sourceFile.Name));
+
+        if (!existingMediaGroupDirectory.Exists) // Erstelle das Mediengruppenverzeichnis, falls es noch nicht existiert
+        {
+            existingMediaGroupDirectory.Create();
+        }
+
+        var expectedTargetFile = new FileInfo(fileInExistingMediaGroupDirectory.FullName);
+
+        File.Copy(sourceFile.FullName, fileInExistingMediaGroupDirectory.FullName, true); // Kopiere die Datei um eine bestehende Mediengruppe zu simulieren
+
+        var command = new MoveFileToMediaLibraryCommand
+        {
+            FilePath = sourceFile.FullName,
+            MediaLibraryPath = mediaLibraryDirectory.FullName,
+            Logger = new NullLogger<MoveFileToMediaLibraryCommand>()
+        };
+
+        FileMovedToMediaLibraryEventArgs? commandEventArgs = null; // Hilfsvariable für das Erfassen der Event-Informationen
+        command.FileMovedToMediaLibrary += (sender, e) => commandEventArgs = e;
+
+        // Act
+        var result = command.Execute(); // Verschiebe die Datei erneut (in die gerade erstellte Mediengruppe)
+
+        // Assert
+        Assert.IsTrue(result.IsSuccess);
+        Assert.IsTrue(commandEventArgs != null);
+        Assert.AreEqual(expectedTargetFile.FullName, commandEventArgs?.FileInfo.FullName);
+        Assert.IsTrue(commandEventArgs?.FileInfo.Exists);
+        Assert.IsTrue(commandEventArgs?.HasMovedToExistingMediaGroup); // Die Datei wurde in eine bestehende Mediengruppe verschoben
+        Assert.IsTrue(commandEventArgs?.HasTargetFileBeenOverwritten); // Die Datei wurde überschrieben
     }
 }
